agent/pyproject.toml
[project]
name = "thufir-agent"
version = "0.1.0"
description = "Supabase data-retrieval agent powered by an OpenAI-compatible LLM"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.12.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]

